feat(functions): dlsite run の自前 reads を網羅計測（SPR-225 P0②）
Stage 1 で creator recompute storm を解消した後も run 窓の type=QUERY が 17k 規模で
残ったが、集計 Monitoring は DB 単位で web と dlsite 関数の QUERY を分離できず、
~15k が dlsite 起因か web（midnight 同時アクセス）かを確定できなかった。
→ dlsite 関数の自前 QUERY を in-process で網羅計測し、3b（ティア差分）の費用対効果＝
3b に進むかの判断材料を得る。

- creator-sync-metrics.ts → dlsite-read-metrics.ts に一般化・改名（軸2/軸3: creator 専用の
  名前で worksMap も持つのは命名のズレ）。run-scoped カウンタに worksMapQueries /
  worksMapDocsRead を追加。
- dlsite-firestore.ts: getExistingWorksMap（全件 skip-check の `where("productId","in",...)`）の
  クエリ回数と返却 docs 数を記録（type=QUERY 主因候補の分解）。
- エンドポイント: run 末尾の計測ログを恒久キー "dlsite reads計測(run)" に統一（worksMap /
  creator-sync / recompute を網羅）。

挙動変更なし（計測のみ）。次 run の本番ログで dlsite 自前 QUERY を確定し、3b/lastFetchedAt(P0①)
に進むかを判断する。
